% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CWD_VolumeFunctions.R
\name{CWD_2021_Vol_calc}
\alias{CWD_2021_Vol_calc}
\title{Title}
\usage{
CWD_2021_Vol_calc(
  CWD_dat = "./data-raw/EP1162CWDsurvey2020-2021.csv",
  Horiz_dat = "./data-raw/EP1162CWDsurveyTransectLines2020-2021.csv",
  out_carbon_comp = FALSE
)
}
\arguments{
\item{out_carbon_comp}{}
}
\description{
Title
}
