% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CWD_VolumeFunctions.R
\name{cwd_sub_boreal_props}
\alias{cwd_sub_boreal_props}
\title{Title}
\usage{
cwd_sub_boreal_props(CWD_dat, size_thresh = 15, out_carbon_comp = TRUE)
}
\arguments{
\item{out_carbon_comp}{}
}
\description{
Title
}
