% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CWD_VolumeFunctions.R
\name{cwd_sub_boreal_vol}
\alias{cwd_sub_boreal_vol}
\title{General SBS coarse woody debris}
\usage{
cwd_sub_boreal_vol(CWD_dat, size_thresh = FALSE, out_carbon_comp = FALSE)
}
\arguments{
\item{size_thresh}{default to FALSE, if pass a number, that is used as the cutoff in size class}

\item{out_carbon_comp}{}
}
\description{
Bruce Rogers data from the SBS - original file name "Erica-SB Plot Data-CWD-Regen". Assuming
2 x 30m transect lengths
}
