% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Trees.R
\name{planted_trees}
\alias{planted_trees}
\title{Number of planted trees per hectare by PSP}
\usage{
planted_trees(planted_data = "./data-raw/Trees/plantedTrees.csv")
}
\arguments{
\item{planted_data}{where is the planted trees data}
}
\description{
Number of planted trees per hectare by PSP
}
\details{
Hybrid spruce (Sx) was filled planted in 14 experimental units to establish Spruce in un- or
poorly stocked stands. These included PSPs 3, 15, and 24, but looking at the data, it looks like
many plots had planted trees

plot area = 0.05 ha
}
