% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Trees.R
\name{plot_sph_size}
\alias{plot_sph_size}
\title{Plot tree summary - by species and size class}
\usage{
plot_sph_size(
  dbh_size_class = 2,
  plot_area = 0.05,
  raw_data = "./data-raw/SummitLakeData.csv"
)
}
\arguments{
\item{raw_data}{}
}
\description{
Plot tree summary - by species and size class
}
