% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processSORTIE.R
\name{rdSortieOutputs}
\alias{rdSortieOutputs}
\title{import and combine sortie outputs}
\usage{
rdSortieOutputs(out_path, run_name, yrs = NULL)
}
\arguments{
\item{yrs}{#might take this away - get it from the data}

\item{Units_to_output}{a vector of character names for which units to include for subplotting outputs}

\item{dist_edge}{\code{\link[=numeric]{numeric()}} how far from unit boundary to allow subplots (in m)}

\item{num_subplots}{\code{\link[=numeric]{numeric()}} how many subplots}

\item{size_subplot}{\code{\link[=numeric]{numeric()}} radius of plot (standard is 7.98m)}

\item{plotting}{TRUE/FALSE - whether or not to display plots with the unit and subplot location}
}
\description{
import and combine sortie outputs
}
\details{
to do: make this more generic where you pass the sortie output (not Date Creek specific) and
subplot based on different patterns
}
