% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CreateSortieInits.R
\name{sl_sortie_inits}
\alias{sl_sortie_inits}
\title{create sortie inits from Summit Lake data}
\usage{
sl_sortie_inits(
  dbh_size_class = 2,
  plot_area = 0.05,
  raw_data = "D:/Github/SummitLakeData/data-raw/SummitLakeData.csv",
  in_dir,
  out_dir
)
}
\description{
initial tree conditions created from 1992 Summit Lake data. This code also integrates the density
of planted trees surveyed in 2021, and backcasts that density to the smallest seedling class
size for the 1992 data. This is likely an underestimate assuming planted seedling mortality from
1992 to 2021, but we also add natural regeneration data and small tree data (<4cm DBH) from
a nearby study in the SBS. In 1992 the smallest tagged diameter size was 4cm, so we use the
Summit Lake data down to that size class.

For every initial condition produced, we draw randomly from a normal distribution of these seedling
densities to create some initial variability in seedling densities between plots. So this will
change with every run of the sl_sortie_inits function.
}
\author{
Alana Clason & Leah Walker
}
