% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CreateSortieInits.R
\name{SORTIEinits}
\alias{SORTIEinits}
\title{Title}
\usage{
SORTIEinits(
  dbh_size_class = 2,
  minDBH = 0,
  maxDBH,
  plot_area = 0.05,
  raw_data = "./data-raw/Trees/SummitLakeData.csv",
  output_dir
)
}
\arguments{
\item{output_dir}{}
}
\description{
Title
}
