% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CWD_VolumeFunctions.R
\name{CWD_2021_Vol_calc}
\alias{CWD_2021_Vol_calc}
\title{2020/2021 CWD volume calculation}
\usage{
CWD_2021_Vol_calc(
  CWD_dat = "./data-raw/EP1162CWDsurvey2020-2021.csv",
  Horiz_dat = "./data-raw/EP1162CWDsurveyTransectLines2020-2021.csv",
  out_carbon_comp = FALSE
)
}
\arguments{
\item{CWD_dat}{}
}
\description{
2020/2021 CWD volume calculation
}
\details{
Calculate Volume using the British Columbia Ministry of Forests and Range (2009) formula
CWD volume (m3/ha) = pi^2/8L  *  sum\link{D2/cos (A)}
Where: 	L = length of total transect (horizontal distance (HD) in m)
HD = SD / Square root of \link{1 + (\% slope / 100)2}
slope was measured - so need to update this equation
D = diameter of each piece of CWD (cm)
A = tilt angle from horizontal for each piece (degrees)
Because tilt angle was not measured in 1992 and 1993, it was assumed to be zero,
so that cos (A) = 1 for all pieces in those years.
}
