% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CWD_VolumeFunctions.R
\name{CWD_vol_calc}
\alias{CWD_vol_calc}
\title{Summit Lakes cwd data - 2020-2021}
\usage{
CWD_vol_calc(dat_loc, incl_sp_decay = FALSE)
}
\arguments{
\item{dat_loc}{what is the directory where all data is stored}
}
\description{
Summit Lakes cwd data - 2020-2021
}
\details{
Summit Lake currently has only one year of CWD, so only call 2021, but can add more
as re-measurements occur
}
